+++ b2sdk/sync/policy_manager.py
@@ -9,9 +9,11 @@
 ######################################################################
 
 from ..scan.path import AbstractPath
-from .policy import CopyAndDeletePolicy, CopyAndKeepDaysPolicy, CopyPolicy, \
-    DownAndDeletePolicy, DownAndKeepDaysPolicy, DownPolicy, UpAndDeletePolicy, \
-    UpAndKeepDaysPolicy, UpPolicy
+from ..transfer.outbound.upload_source import UploadMode
+from .policy import AbstractFileSyncPolicy, CompareVersionMode, CopyAndDeletePolicy, \
+    CopyAndKeepDaysPolicy, CopyPolicy, DownAndDeletePolicy, DownAndKeepDaysPolicy, \
+    DownPolicy, NewerFileSyncMode, UpAndDeletePolicy, UpAndKeepDaysPolicy, UpPolicy
+from .encryption_provider import AbstractSyncEncryptionSettingsProvider
 
 
 class SyncPolicyManager:
@@ -25,19 +27,21 @@ class SyncPolicyManager:
 
     def get_policy(
         self,
-        sync_type,
+        sync_type: str,
         source_path: AbstractPath,
-        source_folder,
+        source_folder: str,
         dest_path: AbstractPath,
-        dest_folder,
-        now_millis,
-        delete,
-        keep_days,
-        newer_file_mode,
-        compare_threshold,
-        compare_version_mode,
-        encryption_settings_provider,
-    ):
+        dest_folder: str,
+        now_millis: int,
+        delete: bool,
+        keep_days: int,
+        newer_file_mode: NewerFileSyncMode,
+        compare_threshold: int,
+        compare_version_mode: CompareVersionMode,
+        encryption_settings_provider: AbstractSyncEncryptionSettingsProvider,
+        upload_mode: UploadMode,
+        absolute_minimum_part_size: int,
+    ) -> AbstractFileSyncPolicy:
         """
         Return a policy object.
 
@@ -53,6 +57,8 @@ class SyncPolicyManager:
         :param int compare_threshold: difference between file modification time or file size
         :param b2sdk.v2.CompareVersionMode compare_version_mode: setting which determines how to compare source and destination files
         :param b2sdk.v2.AbstractSyncEncryptionSettingsProvider encryption_settings_provider: an object which decides which encryption to use (if any)
+        :param b2sdk.v2.UploadMode upload_mode: determines how file uploads are handled
+        :param int absolute_minimum_part_size: minimum file part size for large files
         :return: a policy object
         """
         policy_class = self.get_policy_class(sync_type, delete, keep_days)
@@ -67,6 +73,8 @@ class SyncPolicyManager:
             compare_threshold,
             compare_version_mode,
             encryption_settings_provider,
+            upload_mode,
+            absolute_minimum_part_size,
         )
 
     def get_policy_class(self, sync_type, delete, keep_days):