+++ b2sdk/transfer/outbound/upload_source.py
@@ -10,15 +10,28 @@
 
 import hashlib
 import io
+import logging
 import os
 
 from abc import abstractmethod
 from typing import Optional
+from enum import Enum, unique
 
 from b2sdk.exception import InvalidUploadSource
+from b2sdk.http_constants import DEFAULT_MIN_PART_SIZE
 from b2sdk.stream.range import RangeOfInputStream, wrap_with_range
+from b2sdk.transfer.outbound.copy_source import CopySource
 from b2sdk.transfer.outbound.outbound_source import OutboundTransferSource
-from b2sdk.utils import hex_sha1_of_stream, hex_sha1_of_unlimited_stream, Sha1HexDigest
+from b2sdk.utils import hex_sha1_of_stream, hex_sha1_of_unlimited_stream, Sha1HexDigest, update_digest_from_stream
+
+logger = logging.getLogger(__name__)
+
+
+@unique
+class UploadMode(Enum):
+    """ Mode of file uploads """
+    FULL = 0  #: always upload the whole file
+    INCREMENTAL = 1  #: use incremental uploads when possible
 
 
 class AbstractUploadSource(OutboundTransferSource):
@@ -81,13 +94,14 @@ class UploadSourceBytes(AbstractUploadSource):
         return self.content_sha1 is not None
 
 
-class UploadSourceLocalFile(AbstractUploadSource):
+class UploadSourceLocalFileBase(AbstractUploadSource):
     def __init__(self, local_path, content_sha1=None):
         self.local_path = local_path
         self.content_length = 0
-        self.check_path_and_get_size()
-
         self.content_sha1 = content_sha1
+        self.digest = None
+        self.digest_progress = 0
+        self.check_path_and_get_size()
 
     def check_path_and_get_size(self):
         if not os.path.isfile(self.local_path):
@@ -109,23 +123,40 @@ class UploadSourceLocalFile(AbstractUploadSource):
     def get_content_length(self):
         return self.content_length
 
+    def _get_hexdigest(self, length=0):
+        length = length or self.content_length
+
+        if self.digest is None:
+            self.digest = hashlib.sha1()
+            self.digest_progress = 0
+
+        if length < self.digest_progress:
+            raise ValueError("Length value can not decrease between calls")
+
+        if length > self.digest_progress:
+            with self.open() as fp:
+                range_length = length - self.digest_progress
+                range_ = wrap_with_range(
+                    fp, self.content_length, self.digest_progress, range_length
+                )
+                update_digest_from_stream(self.digest, range_, range_length)
+            self.digest_progress = length
+
+        return self.digest.hexdigest()
+
     def get_content_sha1(self):
         if self.content_sha1 is None:
-            self.content_sha1 = self._hex_sha1_of_file(self.local_path)
+            self.content_sha1 = self._get_hexdigest()
         return self.content_sha1
 
     def open(self):
         return io.open(self.local_path, 'rb')
 
-    def _hex_sha1_of_file(self, local_path):
-        with self.open() as f:
-            return hex_sha1_of_stream(f, self.content_length)
-
     def is_sha1_known(self):
         return self.content_sha1 is not None
 
 
-class UploadSourceLocalFileRange(UploadSourceLocalFile):
+class UploadSourceLocalFileRange(UploadSourceLocalFileBase):
     def __init__(self, local_path, content_sha1=None, offset=0, length=None):
         super(UploadSourceLocalFileRange, self).__init__(local_path, content_sha1)
         self.file_size = self.content_length
@@ -153,6 +184,72 @@ class UploadSourceLocalFileRange(UploadSourceLocalFile):
     def open(self):
         fp = super(UploadSourceLocalFileRange, self).open()
         return wrap_with_range(fp, self.file_size, self.offset, self.content_length)
+
+
+class UploadSourceLocalFile(UploadSourceLocalFileBase):
+    def get_incremental_sources(self, file_version, min_part_size=None):
+        """
+        Split the upload into a copy and upload source constructing an incremental upload
+
+        This will return a list of upload sources.  If the upload cannot split, the method will return [self].
+        """
+
+        if not file_version:
+            logger.debug(
+                "Fallback to full upload for %s -- no matching file on server", self.local_path
+            )
+            return [self]
+
+        min_part_size = min_part_size or DEFAULT_MIN_PART_SIZE
+        if file_version.size < min_part_size:
+            # existing file size below minimal large file part size
+            logger.debug(
+                "Fallback to full upload for %s -- remote file is smaller than %i bytes",
+                self.local_path, min_part_size
+            )
+            return [self]
+
+        if self.get_content_length() < file_version.size:
+            logger.debug(
+                "Fallback to full upload for %s -- local file is smaller than remote",
+                self.local_path
+            )
+            return [self]
+
+        content_sha1 = file_version.get_content_sha1()
+
+        if not content_sha1:
+            logger.debug(
+                "Fallback to full upload for %s -- remote file content SHA1 unknown",
+                self.local_path
+            )
+            return [self]
+
+        if self._get_hexdigest(file_version.size) != content_sha1:
+            logger.debug(
+                "Fallback to full upload for %s -- content in common range differs", self.local_path
+            )
+            return [self]
+
+        logger.debug("Incremental upload of %s is possible.", self.local_path)
+
+        if file_version.server_side_encryption and file_version.server_side_encryption.is_unknown():
+            source_encryption = None
+        else:
+            source_encryption = file_version.server_side_encryption
+
+        sources = [
+            CopySource(
+                file_version.id_,
+                offset=0,
+                length=file_version.size,
+                encryption=source_encryption,
+                source_file_info=file_version.file_info,
+                source_content_type=file_version.content_type,
+            ),
+            UploadSourceLocalFileRange(self.local_path, offset=file_version.size),
+        ]
+        return sources
 
 
 class UploadSourceStream(AbstractUploadSource):