+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Authorizing a key for a single bucket ensures that this bucket is cached
 * `Bucket.ls` operation supports wildcard matching strings
 * Add `large_file_sha1` support
+* Add support for incremental upload and sync
 
 ### Infrastructure
 * Additional tests for listing files/versions